+++ packages/larva-js/src/larva-ui/index.js
@@ -1,12 +1,22 @@
 import reflowForSideSkinAd from '../interface/SideSkinAd/SideSkinAd';
-import initLarvaUiButton from './LarvaUiButton';
+import LarvaUiToggle from './LarvaUiToggle';
 
 import '../vendor/prism';
 
 export default function initLarvaUi() {
-
+	// This rule is meant for React components.
+	// eslint-disable-next-line @wordpress/no-global-event-listener
 	document.addEventListener( 'DOMContentLoaded', () => {
-		initLarvaUiButton();
+		/**
+		 * Initialize local storage toggle button.
+		 */
+		const buttons = [
+			...document.querySelectorAll( '.js-LarvaUiToggle-button' ),
+		];
+
+		buttons.forEach(
+			( el ) => ( el.pmcLarvaUiToggle = new LarvaUiToggle( el ) )
+		);
 
 		/**
 		 * Side Skins